Added mediator met plot

+++ runScanMG5.py
@@ -1,15 +1,14 @@
 #!/usr/bin/env python3
 
-#Uses an input SLHA file to compute cross-sections using MadGraph and the UFO model files
-#The calculation goes through the following steps
 # 1) Run MadGraph using the options set in the input file 
 # (the proc_card.dat, parameter_card.dat and run_card.dat...).
 
 from __future__ import print_function
-import sys,os
+import sys,os,glob
 from configParserWrapper import ConfigParserExt
 import logging,shutil
 import subprocess
+import multiprocessing
 import tempfile
 import time,datetime
 
@@ -24,14 +23,14 @@ def generateProcess(parser):
     model, since it is independent of the 
     numerical values of the model parameters.
     
-    :param parser: ConfigParser object with all the parameters needed
+    :param parser: Dictionary with parser sections.
     
     :return: True if successful. Otherwise False.
     """
     
     
     #Get run folder:    
-    pars = parser.toDict(raw=False)["MadGraphPars"]
+    pars = parser["MadGraphPars"]
     processCard = os.path.abspath(pars["proccard"])    
     if not os.path.isfile(processCard):
         logger.error("Process card %s not found" %processCard)
@@ -58,7 +57,7 @@ def generateProcess(parser):
             f.write(l)
     
     #Generate process
-    mg5Folder = os.path.abspath('../MG5')
+    mg5Folder = os.path.abspath('./MG5')
     run = subprocess.Popen('./bin/mg5_aMC -f %s' %procCard,shell=True,
                                 stdout=subprocess.PIPE,stderr=subprocess.PIPE,
                                 cwd=mg5Folder)
@@ -104,42 +103,58 @@ def getInfoFromOutput(outputStr):
                'cross-section (pb)' : xsec, 'Number of events' : int(nevts)}
     return runInfo
 
-
 def generateEvents(parser):
     
     """
     Runs the madgraph event generation.
     
-    :param parser: ConfigParser object with all the parameters needed
+    :param parser: Dictionary with parser sections.
     
     :return: True if successful. Otherwise False.
     """
 
-    t0 = time.time()
     
-    pars = parser.toDict(raw=False)["MadGraphPars"]
+    t0 = time.time()
+    pars = parser["MadGraphPars"]
+    if not 'runFolder' in pars:
+        logger.error('Run folder not defined.')
+        return False        
+    else:
+        runFolder = pars['runFolder']
+
     if not 'processFolder' in pars:
         logger.error('Process folder not defined.')
         return False        
     else:
         processFolder = pars['processFolder']
+        if not os.path.isdir(processFolder):
+            logger.error('Process folder %s not found.' %processFolder)
+            return False            
+
+    # If run folder does not exist, create it using processFolder as a template:
+    if not os.path.isdir(runFolder):
+        runFolder = shutil.copytree(processFolder,runFolder,
+                                    ignore=shutil.ignore_patterns('Events','*.lhe'),
+                                    symlinks=True)
+        os.makedirs(os.path.join(runFolder,'Events'))       
+        logger.info("Created temporary folder %s" %runFolder) 
             
-    if not os.path.isdir(processFolder):
-        logger.error('Process folder %s not found.' %processFolder)
+    if not os.path.isdir(runFolder):
+        logger.error('Run folder %s not found.' %runFolder)
         return False
 
     if 'runcard' in pars and os.path.isfile(pars['runcard']):    
-        shutil.copyfile(pars['runcard'],os.path.join(processFolder,'Cards/run_card.dat'))
+        shutil.copyfile(pars['runcard'],os.path.join(runFolder,'Cards/run_card.dat'))
     if 'paramcard' in pars and os.path.isfile(pars['paramcard']):
-        shutil.copyfile(pars['paramcard'],os.path.join(processFolder,'Cards/param_card.dat'))    
+        shutil.copyfile(pars['paramcard'],os.path.join(runFolder,'Cards/param_card.dat'))    
 
 
     # By default do not run Pythia or Delphes
-    runPythia = parser.get('options','runPythia')
-    runDelphes = parser.get('options','runDelphes')
+    runPythia = parser['options']['runPythia']
+    runDelphes = parser['options']['runDelphes']
 
-    pythia8File = os.path.join(processFolder,'Cards/pythia8_card.dat')
-    delphesFile = os.path.join(processFolder,'Cards/delphes_card.dat')
+    pythia8File = os.path.join(runFolder,'Cards/pythia8_card.dat')
+    delphesFile = os.path.join(runFolder,'Cards/delphes_card.dat')
     if runDelphes and 'delphescard' in pars:
         if os.path.isfile(pars['delphescard']):
             shutil.copyfile(pars['delphescard'],delphesFile)
@@ -147,9 +162,11 @@ def generateEvents(parser):
     if runPythia and 'pythia8card' in pars:        
         if os.path.isfile(pars['pythia8card']):
             shutil.copyfile(pars['pythia8card'],pythia8File) 
+
+    cleanOutput = parser['options']['cleanOutput']
     
     #Generate commands file:       
-    commandsFile = tempfile.mkstemp(suffix='.txt', prefix='MG5_commands_', dir=processFolder)
+    commandsFile = tempfile.mkstemp(suffix='.txt', prefix='MG5_commands_', dir=runFolder)
     os.close(commandsFile[0])
     commandsFileF = open(commandsFile[1],'w')
     if runPythia:
@@ -162,11 +179,13 @@ def generateEvents(parser):
         commandsFileF.write('detector=OFF\n')
 
     commandsFileF.write('done\n')
-    comms = parser.toDict(raw=False)["MadGraphSet"]
-    #Set a low number of events, since it does not affect the total cross-section value
-    #(can be overridden by the user, if the user defines a different number in the input card)
+    comms = parser["MadGraphSet"]
+    # Set the MadGraph parameters defined in the ini file
     for key,val in comms.items():
         commandsFileF.write('set %s %s\n' %(key,val))
+    # If cleanOutput = True, do not store the HepMC file (faster run):
+    if cleanOutput:
+        commandsFileF.write('set HEPMCoutput:file hepmcremove\n')
 
     #Done setting up options
     commandsFileF.write('done\n')
@@ -174,37 +193,74 @@ def generateEvents(parser):
 
     commandsFileF.close()
     commandsFile = commandsFile[1]      
+
+    ncore = parser['options']['ncore']
     
     logger.info("Generating MG5 events with command file %s" %commandsFile)
-    run = subprocess.Popen('./bin/generate_events < %s' %(commandsFile),
+    run = subprocess.Popen('./bin/generate_events --multicore --nb_core=%i < %s' %(ncore,commandsFile),
                            shell=True,stdout=subprocess.PIPE,
-                           stderr=subprocess.PIPE,cwd=processFolder)
+                           stderr=subprocess.PIPE,cwd=runFolder)
       
     output,errorMsg= run.communicate()
+    runInfo = {'time (s)' : time.time()-t0}
+    runInfo.update(pars)
     # Try to get info from output
     try:
-        runInfo  = getInfoFromOutput(output)
+        runInfo.update(getInfoFromOutput(output))
     except:
-        runInfo = None
+        pass
 
     logger.debug('MG5 event error:\n %s \n' %errorMsg)
     logger.debug('MG5 event output:\n %s \n' %output)
       
-    logger.info("Finished event generation in %1.2f min" %((time.time()-t0)/60.))
     os.remove(commandsFile)
 
-    cleanOutput = parser.get('options','cleanOutput')
-    if cleanOutput and runInfo:
-        lheFile = os.path.join(processFolder,'Events',runInfo['run number'],'unweighted_events.lhe.gz')
+    
+    if cleanOutput and runInfo and 'run number' in runInfo:
+        lheFile = os.path.join(runFolder,'Events',runInfo['run number'],'unweighted_events.lhe.gz')
         logger.debug('Removing  %s' %lheFile)
         if os.path.isfile(lheFile):
             os.remove(lheFile)
-        hepmcFile = os.path.join(processFolder,'Events',runInfo['run number'], '%s_pythia8_events.hepmc.gz'  %runInfo['run tag'])
+        hepmcFile = os.path.join(runFolder,'Events',runInfo['run number'], '%s_pythia8_events.hepmc.gz'  %runInfo['run tag'])
         logger.debug('Removing  %s' %hepmcFile)
         if os.path.isfile(hepmcFile):
             os.remove(hepmcFile)
+        logFile = os.path.join(runFolder,'Events',runInfo['run number'], '%s_pythia8.log'  %runInfo['run tag'])
+        logger.debug('Removing  %s' %logFile)
+        if os.path.isfile(logFile):
+            os.remove(logFile)
 
     return runInfo
+
+def moveFolders(runInfo):
+    """
+    Move the run folders from the temporary running folder
+    to the process folder.
+    """
+
+    logger.info('Finished event generation for run %i in %1.2f min' 
+                %(int(runInfo['runNumber']),runInfo['time (s)']/60.))
+
+    # Get run folder:
+    runFolder = os.path.abspath(runInfo['runFolder'])
+    runNumber = int(runInfo['runNumber'])
+    processFolder = os.path.abspath(runInfo['processFolder'])
+    # If run folder and process folder are the same, do nothing
+    if runFolder == processFolder:
+        return
+    
+    # Move run folder results to process folder
+    eventFolder = os.path.join(processFolder,'Events')
+    runDirs = list(glob.glob(os.path.join(runFolder,'Events','run_*')))
+    if len(runDirs) != 1:
+        logger.error('Something went wrong. Found %i run folders in %s' %(len(runDirs),runFolder))
+        return False
+    runDir = runDirs[0]
+    finalRunDir = os.path.join(eventFolder,'run_%02d' %runNumber)
+    logger.info('Moving %s to %s' %(runDir,finalRunDir))
+    shutil.move(runDir,finalRunDir)
+    logger.info('Deleting temporary folder %s' %(runFolder))
+    shutil.rmtree(runFolder)
 
 
 def main(parfile,verbose):
@@ -227,21 +283,59 @@ def main(parfile,verbose):
     #Get a list of parsers (in case loops have been defined)    
     parserList = parser.expandLoops()
 
+    # Start multiprocessing pool
+    ncpus = -1
+    if parser.has_option("options","ncpu"):
+        ncpus = int(parser.get("options","ncpu"))
+    if ncpus  < 0:
+        ncpus =  multiprocessing.cpu_count()
+    pool = multiprocessing.Pool(processes=ncpus)
+    if ncpus > len(parserList):
+        ncpus = len(parserList)
+    if ncpus > 1:
+        logger.info('Running in parallel with %i processes' %ncpus)
+    else:
+        logger.info('Running in series with a single process')
+
     now = datetime.datetime.now()
-    for i,newParser in enumerate(parserList):
+    children = []
+    for irun,newParser in enumerate(parserList):
         processFolder = newParser.get('MadGraphPars','processFolder')
-        runInfo = None
+        processFolder = os.path.abspath(processFolder)
+        if processFolder[-1] == '/':
+            processFolder = processFolder[:-1]
+        if not os.path.isdir(processFolder):
+            logger.info('Folder %s not found. Running MG5 to create folder.' %processFolder)
+            generateProcess(newParser)
 
-        runMG = newParser.get('options','runMadGraph')
-        if runMG:
-            if not os.path.isdir(processFolder):
-                logger.info('Folder %s not found. Running MG5 to create folder.' %processFolder)
-                generateProcess(newParser)        
-            runInfo = generateEvents(newParser)
+        # Get largest existing events folder:
+        run0 = 1
+        eventsFolder = os.path.join(processFolder,'Events')
+        if os.path.isdir(eventsFolder):
+            for runF in glob.glob(os.path.join(eventsFolder,'run*')):
+                run0 = max(run0,int(os.path.basename(runF).replace('run_',''))+1)
 
-        logger.info("Finished run %i/%i at %s" %(i+1,len(parserList),now.strftime("%Y-%m-%d %H:%M")))
+        # Create temporary folder names if running in parallel
+        if ncpus > 1:
+            # Create temporary folders
+            runFolder = tempfile.mkdtemp(prefix='%s_'%(processFolder),suffix='_run_%02d' %(run0+irun))
+            os.removedirs(runFolder)
+        else:
+            runFolder = processFolder
 
+        newParser.set('MadGraphPars','runFolder',runFolder)
+        newParser.set('MadGraphPars','runNumber','%02d' %(run0+irun))
+
+        parserDict = newParser.toDict(raw=False)
+        logger.debug('submitting with pars:\n %s \n' %parserDict)
+        p = pool.apply_async(generateEvents, args=(parserDict,), callback=moveFolders)                       
+        children.append(p)
+
+#     Wait for jobs to finish:
+    output = [p.get() for p in children]
     logger.info("Finished all runs (%i) at %s" %(len(parserList),now.strftime("%Y-%m-%d %H:%M")))
+
+    return output
